+++ CHANGELOG.rst
@@ -21,12 +21,14 @@ ouster_client
   even when config params have not changed.
 * breaking change: drop defaults parameters from the shortform ``init_client()`` method.
 * bugfix: fix bug in set_config for config_flag auto
+* added a new method ``init_logger()`` to provide control over the logs emitted by ``ouster_client``.
 
 python
 ------
 * breaking change: drop defaults parameters of ``client.Sensor`` constructor.
 * breaking change: change Scans interface Timeout to default to 1 second instead of None
 * bugfix: fix bug in set_config for auto udp destination configuration parameter
+* added a new method ``init_logger()`` to provide control over the logs emitted by ``client.Sensor``.
 
 ouster_viz
 ----------